+++ src/Dagger.jl
@@ -64,7 +64,9 @@ include("table/operations.jl")
 include("table/groupby.jl")
 include("table/join_interface.jl")
 include("table/join.jl")
-include("table/z_dfinterface.jl")
+include("table/dtable_column.jl")
+include("table/dataframes_interface_utils.jl")
+include("table/dataframes_interface.jl")
 
 include("lib/logging-extras.jl")
